+++ src/main/java/com/flint/sample_be_springboot/service/SubjectMasterServiceImpl.java
@@ -1,18 +1,20 @@
 package com.flint.sample_be_springboot.service;
 
 import com.flint.sample_be_springboot.dto.SubjectMasterDTO;
+import com.flint.sample_be_springboot.entity.ClassSubjectAllocationEntity;
 import com.flint.sample_be_springboot.entity.SubjectMasterEntity;
+import com.flint.sample_be_springboot.entity.TeacherClassSubjectAllocationEntity;
 import com.flint.sample_be_springboot.exception.CustomException;
+import com.flint.sample_be_springboot.repository.ClassSubjectAllocationRepository;
 import com.flint.sample_be_springboot.repository.SubjectMasterRepository;
+import com.flint.sample_be_springboot.repository.TeacherClassSubjectAllocationRepository;
 import com.flint.sample_be_springboot.util.BaseService;
 import com.flint.sample_be_springboot.util.CustomQuerySpecification;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
@@ -30,6 +32,12 @@ public class SubjectMasterServiceImpl extends BaseService implements SubjectMast
 
     @Autowired
     private SubjectMasterRepository subjectMasterRepository;
+
+    @Autowired
+    private ClassSubjectAllocationRepository classSubjectAllocationRepository;
+
+    @Autowired
+    private TeacherClassSubjectAllocationRepository teacherClassSubjectAllocationRepository;
 
     @Override
     public SubjectMasterDTO getSubjectMasterById(Long subjectMasterId) {
@@ -49,7 +57,7 @@ public class SubjectMasterServiceImpl extends BaseService implements SubjectMast
         log.info("Enter into saveSubjectMaster");
 
         Optional<SubjectMasterEntity> existingSubject = subjectMasterRepository.findBySubjectName(subjectMasterDTO.getSubjectName());
-        if(existingSubject.isPresent()){
+        if (existingSubject.isPresent()) {
             throw new CustomException("Subject already exist", HttpStatus.PRECONDITION_FAILED);
         }
 
@@ -71,8 +79,8 @@ public class SubjectMasterServiceImpl extends BaseService implements SubjectMast
                 .orElseThrow(() -> new CustomException("Subject information not found", HttpStatus.NOT_FOUND));
 
         Optional<SubjectMasterEntity> isDuplicate = subjectMasterRepository.findBySubjectNameAndSubjectMasterIdNot
-                        (subjectMasterDTO.getSubjectName(), subjectMasterDTO.getSubjectMasterId());
-        if(isDuplicate.isPresent()){
+                (subjectMasterDTO.getSubjectName(), subjectMasterDTO.getSubjectMasterId());
+        if (isDuplicate.isPresent()) {
             throw new CustomException("Subject already exist", HttpStatus.PRECONDITION_FAILED);
         }
 
@@ -92,6 +100,18 @@ public class SubjectMasterServiceImpl extends BaseService implements SubjectMast
 
         SubjectMasterEntity existingSubject = subjectMasterRepository.findById(subjectMasterId)
                 .orElseThrow(() -> new CustomException("Subject information not found", HttpStatus.NOT_FOUND));
+
+        List<ClassSubjectAllocationEntity> classSubjectAllocationEntities = classSubjectAllocationRepository
+                .findBySubjectMasterEntity_SubjectMasterId(existingSubject.getSubjectMasterId());
+        if (classSubjectAllocationEntities != null && !classSubjectAllocationEntities.isEmpty()) {
+            throw new CustomException("This Subject is assigned in subject assignment and teacher subjects, can't delete this ", HttpStatus.FOUND);
+        }
+
+        List<TeacherClassSubjectAllocationEntity> teacherClassSubjectAllocationEntities = teacherClassSubjectAllocationRepository
+                .findBySubjectMasterEntity_SubjectMasterId(existingSubject.getSubjectMasterId());
+        if (teacherClassSubjectAllocationEntities != null && !teacherClassSubjectAllocationEntities.isEmpty()) {
+            throw new CustomException("This Subject is assigned in subject assignment and teacher subjects, can't delete this user", HttpStatus.FOUND);
+        }
 
         subjectMasterRepository.delete(existingSubject);
